+++ llvm/lib/IR/AsmWriter.cpp
@@ -107,6 +107,10 @@ static cl::opt<bool> PreserveAssemblyUseListOrder(
     "preserve-ll-uselistorder", cl::Hidden, cl::init(false),
     cl::desc("Preserve use-list order when writing LLVM assembly."));
 
+static cl::opt<bool> PrintFPMemoryEffects(
+    "print-fp-memory-effects", cl::Hidden,
+    cl::desc("Pretty print floating-point memory effects when dumping"));
+
 // Make virtual table appear in this compilation unit.
 AssemblyAnnotationWriter::~AssemblyAnnotationWriter() = default;
 
@@ -4320,6 +4324,25 @@ void AssemblyWriter::printGCRelocateComment(const GCRelocateInst &Relocate) {
   Out << ")";
 }
 
+static void printFPMemoryEffects(raw_ostream &Out, MemoryEffects ME) {
+  ModRefInfo MR = ME.getModRef(IRMemLocation::InaccessibleMem);
+  Out << " ; fpe=[";
+  switch (MR) {
+  case ModRefInfo::NoModRef:
+    break;
+  case ModRefInfo::Ref:
+    Out << "r";
+    break;
+  case ModRefInfo::Mod:
+    Out << "w";
+    break;
+  case ModRefInfo::ModRef:
+    Out << "rw";
+    break;
+  }
+  Out << "]";
+}
+
 /// printInfoComment - Print a little comment after the instruction indicating
 /// which slot it occupies.
 void AssemblyWriter::printInfoComment(const Value &V, bool isMaterializable) {
@@ -4348,6 +4371,15 @@ void AssemblyWriter::printInfoComment(const Value &V, bool isMaterializable) {
 
   if (PrintInstAddrs)
     Out << " ; " << &V;
+
+  if (PrintFPMemoryEffects) {
+    if (auto *CI = dyn_cast<CallInst>(&V))
+      if (Intrinsic::ID IID = CI->getIntrinsicID())
+        if (const Function *F = CI->getFunction())
+          if (IntrinsicInst::isFloatingPointOperation(IID) &&
+              F->hasFnAttribute(Attribute::StrictFP))
+            printFPMemoryEffects(Out, CI->getMemoryEffects());
+  }
 }
 
 static void maybePrintCallAddrSpace(const Value *Operand, const Instruction *I,