+++ llvm/lib/IR/IRBuilder.cpp
@@ -113,6 +113,73 @@ IRBuilderBase::createCallHelper(Function *Callee, ArrayRef<Value *> Ops,
   return CI;
 }
 
+CallInst *IRBuilderBase::CreateCall(FunctionType *FTy, Value *Callee,
+                                    ArrayRef<Value *> Args,
+                                    ArrayRef<OperandBundleDef> OpBundles,
+                                    const Twine &Name, MDNode *FPMathTag) {
+  // Operand bundles can be extended by floating-point bundles. In this case
+  // we need a copy of OpBundles, because ArrayRef is immutable.
+  ArrayRef<OperandBundleDef> ActualBundlesRef = OpBundles;
+  SmallVector<OperandBundleDef, 2> ActualBundles;
+
+  bool NeedUpdateMemoryEffects = false;
+  if (const auto *Func = dyn_cast<Function>(Callee))
+    if (Intrinsic::ID ID = Func->getIntrinsicID())
+      if (IntrinsicInst::isFloatingPointOperation(ID)) {
+        // If the builder specifies non-default floating-point options, add
+        // corresponding operand bundle unless a bundle with such tag is already
+        // present.
+        bool NeedRounding;
+        bool NeedExceptions;
+        if (IsFPConstrained) {
+          NeedRounding = DefaultConstrainedRounding != RoundingMode::Dynamic;
+          NeedExceptions = DefaultConstrainedExcept != fp::ebStrict;
+        } else {
+          NeedRounding =
+              DefaultConstrainedRounding != RoundingMode::NearestTiesToEven;
+          NeedExceptions = false;
+          assert(DefaultConstrainedExcept == fp::ebIgnore &&
+                 "FP exception in default mode must be ignored");
+        }
+        // Options specified by bundles have higher precedence.
+        for (const auto &Bundle : OpBundles) {
+          if (NeedRounding && Bundle.getTag() == "fp.round")
+            NeedRounding = false;
+          if (NeedExceptions && Bundle.getTag() == "fp.except")
+            NeedExceptions = false;
+        }
+        if (NeedRounding || NeedExceptions) {
+          ActualBundles.append(OpBundles.begin(), OpBundles.end());
+          if (NeedRounding)
+            createRoundingBundle(ActualBundles, DefaultConstrainedRounding);
+          if (NeedExceptions)
+            createExceptionBundle(ActualBundles, DefaultConstrainedExcept);
+          ActualBundlesRef = ActualBundles;
+        }
+        if (IsFPConstrained) {
+          // Due to potential setting FP exception bits, in modes other than
+          // the default, the memory effects must include read/write access
+          // to FPE.
+          MemoryEffects FME = Func->getMemoryEffects();
+          NeedUpdateMemoryEffects = !FME.doesAccessInaccessibleMem();
+        }
+      }
+
+  // If the call accesses FPE, update memory effects accordingly.
+  CallInst *CI = CallInst::Create(FTy, Callee, Args, ActualBundlesRef);
+  if (NeedUpdateMemoryEffects) {
+    MemoryEffects ME = MemoryEffects::inaccessibleMemOnly();
+    auto A = Attribute::getWithMemoryEffects(getContext(), ME);
+    CI->addFnAttr(A);
+  }
+
+  if (IsFPConstrained)
+    setConstrainedFPCallAttr(CI);
+  if (isa<FPMathOperator>(CI))
+    setFPAttrs(CI, FPMathTag, FMF);
+  return Insert(CI, Name);
+}
+
 static Value *CreateVScaleMultiple(IRBuilderBase &B, Type *Ty, uint64_t Scale) {
   Value *VScale = B.CreateVScale(Ty);
   if (Scale == 1)
@@ -1277,6 +1344,16 @@ CallInst *IRBuilderBase::CreateDereferenceableAssumption(Value *PtrValue,
   OperandBundleDefT<Value *> DereferenceableOpB("dereferenceable", Vals);
   return CreateAssumption(ConstantInt::getTrue(getContext()),
                           {DereferenceableOpB});
+}
+
+void IRBuilderBase::createRoundingBundle(
+    SmallVectorImpl<OperandBundleDef> &Bundles, RoundingMode RM) {
+  addRoundingBundle(Context, Bundles, RM);
+}
+
+void IRBuilderBase::createExceptionBundle(
+    SmallVectorImpl<OperandBundleDef> &Bundles, fp::ExceptionBehavior Except) {
+  addExceptionBundle(Context, Bundles, Except);
 }
 
 IRBuilderDefaultInserter::~IRBuilderDefaultInserter() = default;